ai_coach/requirements.txt
# AI Interview Coach - Dependencies

# Gemini AI for answer evaluation
google-genai>=1.50.0

# Supabase for database access
supabase>=2.0.0

# Python dotenv for environment variables
python-dotenv>=1.0.0

# Optional: For web interface
# streamlit>=1.28.0
# flask>=3.0.0

# Optional: For data analysis
# pandas>=2.0.0
# numpy>=1.24.0
